Feat: Add new setting for controlling the order of labels in compressed proofs
This new setting, yamma.labelsOrderInCompressedProof, allows users to control the order of labels in compressed proofs. The three possible values for the setting are:

fifo: The labels are ordered in the order in which they appear in the packed proof.
mostReferencedFirst: The labels are ordered by the number of occurrences in the packed proof, with the most referenced labels appearing first.
mostReferencedFirstAndNiceFormatting: The labels are ordered by the number of occurrences in the packed proof, with the most referenced labels appearing first. However, the labels are also grouped together based on their index in the packed proof, and the labels within each group are subject to a knapsack algorithm to nicely fit the line width.

The default value for the setting is mostReferencedFirstAndNiceFormatting.

With this commit, the new setting works with the unified proof in .mmp file, but it does not work with .mmt files, yet.